requirements.txt
# requirements.txt
pyyaml==6.0.1
pytest==8.1.1
python-dotenv==1.0.1
groq==1.1.2
streamlit
plotly
pandas